(fix phpstan) set bank account on new membership subscriptions
set bank account information on new membership subscriptions

When creating new membership subscriptions and choosing to create an invoice, existing bank account information should be used. Added

+++ htdocs/adherents/class/adherent.class.php
@@ -1841,7 +1841,7 @@ class Adherent extends CommonObject
 					$invoice->fk_account = $customer->fk_account;
 				} elseif (getDolGlobalString('FACTURE_RIB_NUMBER')) {
 					// set default bank account from invoice module settings
-					$invoice->fk_account = getDolGlobalString('FACTURE_RIB_NUMBER');
+					$invoice->fk_account = (int) getDolGlobalString('FACTURE_RIB_NUMBER');
 				}
 				//$invoice->date = $datesubscription;
 				$invoice->date = dol_now();
